avro-3603 relocate files, update namespaces
