+++ lang/csharp/src/apache/main/Reflect/Converter/FuncFieldConverter.cs
@@ -18,7 +18,7 @@
 
 using System;
 
-namespace Avro.Reflect
+namespace Avro.Reflect.Converter
 {
     /// <summary>
     /// Field converter using a Func
