+++ lang/csharp/src/apache/test/Reflect/TestLogMessage.cs
@@ -21,6 +21,7 @@ using System.Collections.Generic;
 using System.IO;
 using Avro.IO;
 using Avro.Reflect;
+using Avro.Reflect.Conver;
 using NUnit.Framework;
 
 namespace Avro.Test
